+++ CHANGELOG.md
@@ -31,10 +31,11 @@
 * ROCm: rocm-core - `5.4.0.50400-72`
 * Protobuf - [V3.12.4](https://github.com/protocolbuffers/protobuf/releases/tag/v3.12.4)
 * OpenCV - [4.6.0](https://github.com/opencv/opencv/releases/tag/4.6.0)
-* RPP - [0.98](https://github.com/GPUOpen-ProfessionalCompute-Libraries/rpp/releases/tag/0.98)
+* RPP - [1.2.0](https://github.com/GPUOpen-ProfessionalCompute-Libraries/rpp/releases/tag/1.2.0)
 * FFMPEG - [n4.4.2](https://github.com/FFmpeg/FFmpeg/releases/tag/n4.4.2)
+* MIVisionX - [rocm-5.6.0](https://github.com/GPUOpen-ProfessionalCompute-Libraries/MIVisionX/releases/tag/rocm-5.6.0)
 * Dependencies for all the above packages
-* rocAL Setup Script - `V1.0.0`
+* rocAL Setup Script - `V1.0.2`
 
 ### Known issues